feat: Email processing class.
Added the service class MailReader.java in mail-supplier-service with a method to process invoices emails and return their xml attachments as a list of string via IMAP.

test: Email processing class tests.

Added MailReaderIntegrationTest.java to test the connection to the IMAP server of a mail supplier using GreenMail.
Added MailReaderTest.java to test the attachment processing.